# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXIS Risk-IQ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /controls:
    get:
      operationId: listControls
      tags: [controls]
      summary: List controls for a tenant
      parameters:
        - name: tenantId
          in: query
          required: true
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
        - name: withinAppetite
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of controls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Control"
    post:
      operationId: createControl
      tags: [controls]
      summary: Create a new control
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlInput"
      responses:
        "201":
          description: Created control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
  /controls/{controlId}:
    get:
      operationId: getControl
      tags: [controls]
      summary: Get control detail
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Control detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlDetail"
    patch:
      operationId: updateControl
      tags: [controls]
      summary: Perform a named action on a control
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlUpdate"
      responses:
        "200":
          description: Updated control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
    put:
      operationId: updateControlFields
      tags: [controls]
      summary: Update control field values
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlFieldUpdate"
      responses:
        "200":
          description: Updated control
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Control"
    delete:
      operationId: deleteControl
      tags: [controls]
      summary: Delete a control
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /controls/{controlId}/notes:
    post:
      operationId: addControlNote
      tags: [controls]
      summary: Add a note to a control
      parameters:
        - name: controlId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
...
  /tenants/{tenantId}/controls/import:
    post:
      operationId: importControls
      tags: [controls]
      summary: Bulk import controls from client-parsed CSV/Excel
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportControlsInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportControlsResult"
  /tenants/{tenantId}/controls/export:
    get:
      operationId: exportControls
      tags: [controls]
      summary: Export controls as CSV
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    Control:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        department:
          type: string
        departmentId:
          type: string
        risk:
          type: string
        riskId:
          type: string
        control:
          type: string
        controlOwner:
          type: string
        implementationDate:
          type: string
        residualRiskScore:
          type: number
        withinAppetite:
          type: boolean
        overallRiskLevel:
          type: string
        implementationStatus:
          type: string
        lastReviewed:
          type: ["string", "null"]
        noteCount:
          type: integer
        isEscalated:
          type: boolean
        escalatedAt:
          type: ["string", "null"]
      required:
        - id
        - tenantId
        - department
        - departmentId
        - risk
        - riskId
        - control
        - controlOwner
        - implementationDate
        - residualRiskScore
        - withinAppetite
        - overallRiskLevel
        - implementationStatus
        - noteCount
        - isEscalated
    ControlDetail:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        department:
          type: string
        departmentId:
          type: string
        risk:
          type: string
        riskId:
          type: string
        control:
          type: string
        controlOwner:
          type: string
        implementationDate:
          type: string
        residualRiskScore:
          type: number
        withinAppetite:
          type: boolean
        overallRiskLevel:
          type: string
        implementationStatus:
          type: string
        lastReviewed:
          type: ["string", "null"]
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEntry"
        isEscalated:
          type: boolean
        escalatedAt:
          type: ["string", "null"]
        riskDescription:
          type: string
        controlDescription:
          type: string
        inherentRiskScore:
          type: number
      required:
        - id
        - tenantId
        - department
        - departmentId
        - risk
        - riskId
        - control
        - controlOwner
        - implementationDate
        - residualRiskScore
        - withinAppetite
        - overallRiskLevel
        - implementationStatus
        - notes
        - timeline
        - isEscalated
        - riskDescription
        - controlDescription
        - inherentRiskScore
    ControlUpdate:
      type: object
      properties:
        implementationStatus:
          type: string
        controlOwner:
          type: string
        implementationDate:
          type: string
        residualRiskScore:
          type: number
        withinAppetite:
          type: boolean
    NoteInput:
      type: object
      required: [content, author]
      properties:
        content:
          type: string
        author:
          type: string
    Note:
      type: object
      properties:
        id:
          type: string
        controlId:
          type: string
        content:
          type: string
        author:
          type: string
        createdAt:
          type: string
      required: [id, controlId, content, author, createdAt]
    TimelineEntry:
      type: object
      properties:
        id:
          type: string
        action:
          type: string
        actor:
          type: string
        timestamp:
          type: string
        description:
          type: string
        result:
          type: string
      required: [id, action, actor, timestamp, description, result]
...
    ControlInput:
      type: object
      required: [tenantId, departmentId, risk, control]
      properties:
        tenantId:
          type: string
        departmentId:
          type: string
        department:
          type: string
        risk:
          type: string
        riskId:
          type: string
        control:
          type: string
        controlOwner:
          type: string
        implementationDate:
          type: string
        residualRiskScore:
          type: number
        inherentRiskScore:
          type: number
        withinAppetite:
          type: boolean
        overallRiskLevel:
          type: string
          enum: [High, Medium, Low]
        implementationStatus:
          type: string
          enum: [Implemented, In Progress, Overdue, Draft, Awaiting Review, Escalated]
        riskDescription:
          type: string
        controlDescription:
          type: string
        actor:
          type: string
        actorRole:
          type: string
    ControlFieldUpdate:
      type: object
      properties:
        risk:
          type: string
        control:
          type: string
        controlOwner:
          type: string
        implementationDate:
          type: string
        residualRiskScore:
          type: number
        inherentRiskScore:
          type: number
        withinAppetite:
          type: boolean
        overallRiskLevel:
          type: string
        implementationStatus:
          type: string
        riskDescription:
          type: string
        controlDescription:
          type: string
        actor:
          type: string
        actorRole:
          type: string
    ImportControlsInput:
      type: object
      required: [rows]
      properties:
        rows:
          type: array
          items:
            type: object
            required: [risk, control]
            properties:
              department:
                type: string
              risk:
                type: string
              control:
                type: string
              controlOwner:
                type: string
              implementationDate:
                type: string
              overallRiskLevel:
                type: string
              implementationStatus:
                type: string
              withinAppetite:
                type: string
              residualRiskScore:
                type: number
              inherentRiskScore:
                type: number
              riskDescription:
                type: string
              controlDescription:
                type: string
        actor:
          type: string
        actorRole:
          type: string
    ImportControlsResult:
      type: object
      required: [inserted, errors]
      properties:
        inserted:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              error:
                type: string
            required: [row, error]
...