# lib/api-zod/src/generated/types/control.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXIS Risk-IQ API
 * OpenAPI spec version: 0.1.0
 */

export interface Control {
  id: string;
  tenantId: string;
  department: string;
  departmentId: string;
  risk: string;
  riskId: string;
  control: string;
  controlOwner: string;
  implementationDate: string;
  residualRiskScore: number;
  withinAppetite: boolean;
  overallRiskLevel: string;
  implementationStatus: string;
  /** @nullable */
  lastReviewed?: string | null;
  noteCount: number;
  isEscalated: boolean;
  /** @nullable */
  escalatedAt?: string | null;
}