# lib/api-zod/src/generated/types/controlInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXIS Risk-IQ API
 * OpenAPI spec version: 0.1.0
 */
import type { ControlInputImplementationStatus } from './controlInputImplementationStatus';
import type { ControlInputOverallRiskLevel } from './controlInputOverallRiskLevel';

export interface ControlInput {
  tenantId: string;
  departmentId: string;
  department?: string;
  risk: string;
  riskId?: string;
  control: string;
  controlOwner?: string;
  implementationDate?: string;
  residualRiskScore?: number;
  inherentRiskScore?: number;
  withinAppetite?: boolean;
  overallRiskLevel?: ControlInputOverallRiskLevel;
  implementationStatus?: ControlInputImplementationStatus;
  riskDescription?: string;
  controlDescription?: string;
  actor?: string;
  actorRole?: string;
}