# lib/api-zod/src/generated/types/importControlsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NEXIS Risk-IQ API
 * OpenAPI spec version: 0.1.0
 */
import type { ImportControlsResultErrorsItem } from './importControlsResultErrorsItem';

export interface ImportControlsResult {
  inserted: number;
  errors: ImportControlsResultErrorsItem[];
}